🧹 chore(tests): Move to new test file
